requirements.txt
fastapi
uvicorn
selenium
python-dotenv
nest_asyncio